refactor: Add usePopulationLabel and usePrefectureCheck hooks for "Separation of concerns"
